Use 64-bit integers for node and graph x/y.
The graph_x feature for devmap was overflowing.

github.com/ChrisCummins/ProGraML/issues/37

+++ deeplearning/ml4pl/graphs/labelled/dataflow/alias_set/alias_set.py
@@ -145,11 +145,11 @@ def MakeAliasSetGraphs(
   for function, alias_set in function_alias_set_pairs:
     # Translate the must/may alias property into 3-class 1-hot labels.
     if alias_set.type == "may alias":
-      false = np.array([1, 0, 0], np.int32)
-      true = np.array([0, 1, 0], np.int32)
+      false = np.array([1, 0, 0], np.int64)
+      true = np.array([0, 1, 0], np.int64)
     elif alias_set.type == "must alias":
-      false = np.array([1, 0, 0], np.int32)
-      true = np.array([0, 0, 1], np.int32)
+      false = np.array([1, 0, 0], np.int64)
+      true = np.array([0, 0, 1], np.int64)
     else:
       raise ValueError(f"Unknown alias set type `{alias_set.type}`")
